requirements.txt
absl-py==2.1.0
albumentations==1.4.4
annotated-types==0.6.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.0.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deepface==0.0.90
dlib @ file:///C:/Users/thein/Downloads/Dlib-library-Installation-main/Dlib-library-Installation-main/dlib-master/dlib-master/dlib-19.22.99-cp310-cp310-win_amd64.whl
easyocr==1.7.1
exceptiongroup==1.2.0
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.4
fire==0.6.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gdown==5.1.0
google-pasta==0.2.0
grpcio==1.62.1
gunicorn==22.0.0
h5py==3.10.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgviz==1.7.5
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.1.1
kiwisolver==1.4.5
labelme==5.4.1
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.7
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
onnxruntime==1.17.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyclipper==1.3.0.post5
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-version==0.0.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
QtPy==2.4.1
requests==2.31.0
retina-face==0.0.17
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.4.1.post1
scipy==1.13.0
shapely==2.0.4
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
threadpoolctl==3.4.0
tifffile==2024.4.18
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
